% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{autolayer.timeMAXdata}
\alias{autolayer.timeMAXdata}
\title{Autolayer Method for \code{timeMAXdata} Objects}
\usage{
\method{autolayer}{timeMAXdata}(object, which = c("segments", "points"), ...)
}
\arguments{
\item{object}{An object with class \code{"timeMAXdata"} representing
censored observations of a same variables.}

\item{which}{Character vector giving the layer to add: \code{"points"}
shows the non-censored observations as points and \code{"segments"}
shows the censored observations as vertical segements.}

\item{...}{Not used yet.}
}
\value{
A graphical layer to be added to an existing \code{ggplot}
object.
}
\description{
Autolayer method for \code{timeMAXdata} objects representing censored
observations of a same variable in time.
}
