% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEVBayesggplot.R
\name{autoplot.predRLTVGEVBayes}
\alias{autoplot.predRLTVGEVBayes}
\title{Autoplot a \code{PredRLTVGEVBayes} object}
\usage{
\method{autoplot}{predRLTVGEVBayes}(object, ...)
}
\arguments{
\item{object}{A \code{predRLTVGEVBayes} object.}

\item{...}{Not used yet.}
}
\value{
A graphical object inheriting from \code{"ggplot"}.
}
\description{
Autoplot the predictive distribution computed from a
\code{TVGEVBayes} object, i.e. the result returned by the
\code{predict} method for such an object.
}
