% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLikTVGEVBayes.R
\name{autoplot.profLik.TVGEVBayes}
\alias{autoplot.profLik.TVGEVBayes}
\title{Autoplot Method for \code{TVGEVBayes} Objects}
\usage{
\method{autoplot}{profLik.TVGEVBayes}(object, ...)
}
\arguments{
\item{object}{A \code{TVGEVBayes} object.}

\item{...}{Not used.}
}
\value{
An object inheriting from the \code{"ggplot"} class.
}
\description{
This method returns a trellis graphics with one facet
per "parameter". Each facet shows the profile log-likelihood
obtained as the upper envelope of a scatterplot.  For a given
confidence level, one horizontal line is added, corresponding
to the level \eqn{\ell_{\text{max}} - \delta}{ell_max - delta}
of the log-likelihood, where \eqn{\delta :=
    q_{\chi^2(1)}(\gamma) / 2 }{delta := qchisq(\gamma) /2.0} and
\eqn{\gamma} is the confidence level. Several confidence
levels can be used by passing a vector as \code{level}.
}
\details{
The \code{profLik.TVGEVBayes} object given in
\code{object} can correspond to a function with a \emph{vector
value} then defining several "parameters". The plot is then in
"trellis graphics" style, with one facet per parameter.
}
\examples{
example(profLik.TVGEVBayes)
autoplot(prof_q100)
autoplot(prof_RLS)
autoplot(prof_Gumb1)
autoplot(prof_Gumb2)
}
\seealso{
\code{\link{TVGEVBayes}}.
}
