% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{autoplot.timeMAXdata}
\alias{autoplot.timeMAXdata}
\title{Autoplot Method for \code{timeMAXdata} Objects}
\usage{
\method{autoplot}{timeMAXdata}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"timeMAXdata"} representing
censored observations of a same variable.}

\item{...}{Not used yet.}
}
\value{
A graphical object inheriting from \code{"ggplot"}.
}
\description{
Autoplot method for \code{timeMAXdata} objects representing censored
observations of a same variable in time.
}
\examples{
## block maxima style. There should be only one observation by block (year)
M <- timeMAXdata("1961_1980" = list("1961" = 7, "1973" = 40, "1979" = c(3, 30)),
                 "1930_1960" = list("1940" = c(10, Inf), "1943" = 22),
                 "1900_1929" = list("1910" = c(14, Inf), "1929" = c(20, 30))) 
g <- autoplot(M)

}
