% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEVBayesggplot.R
\name{autoplot.TVGEVBayes}
\alias{autoplot.TVGEVBayes}
\title{Autoplot Method for \code{TVGEVBayes} Objects}
\usage{
\method{autoplot}{TVGEVBayes}(object, ...)
}
\arguments{
\item{object}{A \code{TVGEVBayes} object.}

\item{...}{Further arguments passed to
\code{\link{fitted.TVGEVBayes}}, such as \code{wich} or
\code{level}.}
}
\value{
A graphical object inheriting from \code{"ggplot"}.
}
\description{
Autoplot a \code{TVGEVBayes} object, trying to show the dependence
of the GEV marginal distribution on the time/date variable.
}
\seealso{
\code{\link{TVGEVBayes}}.
}
