% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{blockCuts}
\alias{blockCuts}
\title{Define the Time Blocks of a Given Time Range}
\usage{
blockCuts(timeRange, blockDuration = "year")
}
\arguments{
\item{timeRange}{A time range. Can either be a character string to
be passed to the \code{str} argument of the
\code{\link{timeRange}} function, or a \code{Date} object returned
by this function.}

\item{blockDuration}{The duration of the blocks. Should be an integer
number of years as in \code{"2 years"}.}
}
\value{
A vector with class \code{"Date"} giving the cuts
for the blocks. By dropping the last element we get the beginnings
of the blocks and by dropping the first element we get the ends of
the blocks.
}
\description{
Define the cuts (dates) for the time blocks in a given time range.
}
\examples{
tB <- blockCuts("1930_2020", block = "2 years")

}
