% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{doIntersect}
\alias{doIntersect}
\title{Find Possible Intersections among Periods of Time}
\usage{
doIntersect(Periods, details = FALSE)
}
\arguments{
\item{Periods}{Either a character vector with each of its elements
\code{\link{timeRange}}, or a \code{"Date"} matrix with one row
per period and wto colums corresponding to the beginning and the
end.}

\item{details}{Logical. If \code{TRUE} the results are detailed
and tell for all couples of periods if there is intersection or
not.}
}
\value{
A logical with length \code{1} or a data frame with
one row per couple of periods.
}
\description{
Find possible intersections between given periods of time or time
ranges.
}
\examples{
per1 <- timeRange("1990_2000")
per2 <- timeRange("1970_1992")
per3 <- timeRange("1930_1960")
Periods <- rbind(per1, per2, per3)
doIntersect(Periods)
doIntersect(Periods, details = TRUE)
doIntersect(c("1990_2000", "1970_1992", "1930_1960"), details = TRUE)

}
