% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLikTVGEVBayes.R
\name{envProf}
\alias{envProf}
\title{Compute Confidence Limits using Sampled Vectors of
Parameters and Log-Likelihood Values}
\usage{
envProf(param, logLik, level = 0.95, plot = 0)
}
\arguments{
\item{param}{A numeric vector giving the values of the parameter
of interest.}

\item{logLik}{A numeric vector containing the log-likelihood.}

\item{level}{A vector of confidence levels.}

\item{plot}{An integer giving the "level of verbosity" of the
plot. The value \code{0L} corresponds to no plot shown.}
}
\value{
A matrix with two lines and one column by confidence
level. The columns correspond to the confidence levels in
increasing order so the confidence interval becomes larger as
the column index increases.
}
\description{
Compute confidence limits by using a sampled vector
of parameter values \eqn{\psi^{[k]}}{psi[k]} (usually MCMC
iterates) along with the corresponding values
\eqn{\ell(\psi^{[k]})}{ell(psi[k])} of the log-likelihood. For
each given confidence level \eqn{\gamma_i}, the confidence
limits are the values of \eqn{\psi} for which the
log-likelihood \eqn{\ell(\psi)} is equal to
\eqn{\ell_{\text{max}} - \delta_i}{ell_max - delta[i]} where
\eqn{\ell_{\text{max}}}{ell_max} is the maximal log-likelihood
and
\eqn{\delta_i:= q_{\chi^2(1)}(\gamma_i) / 2.0}{delta_i = qchisq(gamma_i, df = 1) / 2.0}.
}
\details{
First, a threshold \eqn{u} for the log-likelihood is
chosen below the smallest \eqn{\ell_{\text{max}} -
    \delta_i}{ell_max - delta[i]}. The points \eqn{[\psi^{[k]},
    \ell^{[k]}]} falling above \eqn{u} are selected and their
convex hull is computed. The upper part of the convex hull
provides the profile curve as a continous broken line. The two
confidence limits are obtained as the interesections of the
profile with the horizontal line. Inasmuch as the points have
a good coverage of the region of the parameter space where the
log-likelhood is high, we get a good approximation ot the
profile likelihood and of the confidence limits.
}
\examples{
df <- within(TXMax_Dijon, Date <- as.Date(sprintf("\%4d-01-01", Year)))
object <- TVGEVBayes(data = df, date = "Date", response = "TXMax",
                     design = polynomX(date = Date, degree = 1),
                     loc = ~ t1, scale = ~ 1, shape = ~ 1,
                    seed = 1234)
MCMC <- apply(object$MCMC, 3, rbind)
logLik <- -apply(MCMC, MARGIN = 1, FUN = object$TVGEV$negLogLikFun,
                 object = object$TVGEV)
## envProf(param= MCMC[ , 1], logLik)
envProf(param= MCMC[ , 1], logLik, level = c(0.70, 0.95))

}
\keyword{internal}
