% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEVBayesMethods.R
\name{newTimeRange}
\alias{newTimeRange}
\title{Default "New" Time Range}
\usage{
newTimeRange(object, timeRange = NULL, ...)
}
\arguments{
\item{object}{A \code{TVGEVBayes} object.}

\item{timeRange}{A character with length one defining the new time
range. See the \code{\link{timeRange}} function.}

\item{...}{Not used yet.}
}
\value{
Alist with the two elements
\item{\code{timeRange} }{ A character with length one
describing the time range as in \code{"2020_2019"}.
}
\item{\code{date} }{ A \code{Date} vector with length 2
giving the beginning and the end of the time-range.
}
}
\description{
Guess a default "new" time range for a \code{TVGEVBayes} object.
}
\details{
Given a \code{TVGEVBayes} object we can find both the block
duration and the beginning and end of the last block. The default
new block is the next block with the same duration. Note that if
\code{object} is based on a block duration of \eqn{w > 2} years,
the underlying model does not know anything about the maxima over
blocks with smaller duration, hence we can not predict the maximum
for a block with duration \eqn{< w}.
}
