% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEVBayesMethods.R
\name{predict.TVGEVBayes}
\alias{predict.TVGEVBayes}
\title{Prediction from a \code{TVGEVBayes} Object}
\usage{
\method{predict}{TVGEVBayes}(
  object,
  newTimeRange = NULL,
  prob,
  type = "RL",
  approx = FALSE,
  trace = 0,
  ...
)
}
\arguments{
\item{object}{A \code{TVGEVBayes} object.}

\item{newTimeRange}{A character describing the "new" time range
(period of time) as in \code{\link{timeRange}}.}

\item{prob}{A vector of exceedance probabilities at which the
predictive distribution is to be computed.}

\item{type}{The type of result wanted. For now only the predictive
Return Levels can be computed.}

\item{approx}{\code{Logical}. If \code{TRUE} the predictive
quantiles will be computed by using a linear interpolation method,
while a series of \code{\link{uniroot}} is used instead.}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\description{
Perform some computations about the predictive distribution from a
\code{TVGEVBayes} object. For a "new" period of time given with
\code{newTimeRange}, we can consider the maximum of the block maxima
on this period. The distribution is not a GEV distribution in general,
and due to the uncertainty on the parameters
}
\examples{
 
newTimeRange <- "2020_2026"

}
