% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLikTVGEVBayes.R
\name{print.profLik.TVGEVBayes}
\alias{print.profLik.TVGEVBayes}
\title{Print Method for \code{profLik.TVGEVBayes} Objects}
\usage{
\method{print}{profLik.TVGEVBayes}(x, ...)
}
\arguments{
\item{x}{The \code{profLik.TVGEVBayes} object, created with the
\code{profLik} method of the class \code{"TVGEVBayes"}.}

\item{...}{Not used.}
}
\value{
Nothing. Print the object.
}
\description{
Print Method for \code{profLik.TVGEVBayes} Objects
}
\seealso{
\code{\link{profLik.TVGEVBayes}},
\code{\link{autoplot.profLik.TVGEVBayes}}.
}
