% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLTVGEVBayes.R
\name{RL}
\alias{RL}
\title{Return Levels}
\usage{
RL(object, ...)
}
\arguments{
\item{object}{An object thar can be used to compute return levels.}

\item{...}{Further arguments for methods.}
}
\value{
A table of return levels, typically given as an object
inheriting from the \code{"data.frame"} class.
}
\description{
Compute return levels in relation with return periods.
}
