% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{timeMAXdata}
\alias{timeMAXdata}
\title{Define MAX Data with Time Stamps}
\usage{
timeMAXdata(..., blocks = FALSE)
}
\arguments{
\item{...}{A \emph{named} collection of objects each being
a list or a numeric vector. See \bold{Examples}.}

\item{blocks}{Logical. If \code{TRUE}, the results contain a
a list of \emph{block time series} corresponding to the periods.
More precisely, each row of the \code{MAXdata} data frame correspond
to a block, usually with a one-year duration.}
}
\value{
An object with class \code{timeMAXdata}. This is mainly a
list with the following two elements.
\item{MAXinfo }{
A data frame with columns \code{period}, \code{start},
\code{end}, \code{duration}. Each row corrrespond to
a period or time range with beginning, end and duration
indicated by the columns.
}
\item{MAXdata }{
A data frame with columns \code{block}, \code{period},
\code{date}, \code{y}, \code{yL}, \code{yU}. The column
\code{block} is for compatibility reason, \code{period} links do
\code{MAXinfo}.  The column \code{date} give the date of an event
in the period or the begining of the year. The columns \code{y},
\code{yL} and \code{yU} give the observed level for the variable
of interest and the bounds when the observation is censored.
}

Unfortunately the name \emph{block} in \code{MAXdata} does not
correspond to a block in the block maxima meaning, but rather to a
period or time range.
}
\description{
Define MAX data with time stamps that can be used with
time-varying extreme-value models.
}
\details{
This data structure is inspired by the \code{MAX} data structure
of \emph{Renext}. The data relates to a certain variable which is
recorded in time but can be censored over periods. The data
describes a number of \emph{periods} in time (or \emph{time
ranges}, see \code{\link{timeRange}}), in general as so-called
\emph{historical period}. For each period, the largest
observations of the variable are given along with a year or date
at which the observation was made. The observations for a period
are given either as a \emph{named} numeric vector or as a
\code{named} list, the names being dates or years in both
cases. When a numeric vector is given, each element of the vector
is the observation for the year of date. When a list is given, a
list element can be either a vector of length \eqn{1} giving the
observation or a vector of length \eqn{2} giving the lower and
upper bounds of an interval in which the observation is known to
lie: this is \emph{interval censoring}. An upper bound can be
\code{Inf}. However a lower bound can not be \code{-Inf} in the
block maxima framework unless the period is a one-block period.
}
\section{Caution}{
 Within a given period, we find observations (I)
for which some information is given explicitly, and observations
(N) for which no information explicitely given. The observations
(N) are assumed to be smaller than all the observations (I), hence
to be smaller than the smallest lower bound provided for the
observations (I), say \code{minyL}. As a result, each observation
(N) is considered as surely lying in the interval \code{(-Inf,
minyL)}. So as a rule, an explicit information \code{(yL, yU)}
should correspond to a quite large value of \code{yL} because this
applies to one of the (few) largest observation in the period.
}

\examples{
## block maxima style. There should be only one observation by block (year)
timeMAXdata("1961_1980" = list("1961" = 7, "1973" = 40, "1979" = c(3, 30)),
            "1930_1960" = list("1940" = c(14, Inf), "1943" = 22),
            "1900_1929" = list("1910" = c(10, Inf), "1929" = c(20, 30))) 

## marked process (POT) style. The observations are attached to a day,
## so there can 
timeMAXdata("1961_1980" = list("1961-12-02" = 7, "1973-12-20" = 40,
                               "1976-01-21" = c(3, 30), "1979-02-19" = 27),
            "1930_1960" = list("1940" = c(14, Inf), "1943" = 22),
            "1900_1929" = list("1910" = c(10, Inf), "1929" = c(12, 30))) 

}
\seealso{
\code{\link{autoplot.timeMAXdata}} and
\code{\link{autolayer.timeMAXdata}}.
}
