% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{timeRange}
\alias{timeRange}
\title{Define a Time Range}
\usage{
timeRange(str)
}
\arguments{
\item{str}{A character string in suitable format, see \bold{Examples}.}
}
\value{
An object with class \code{"Date"} and length 2 containing
the beginning and the end. The time-range or "period" is formed by
all dates which are \code{>=} the beginning and are \code{<} the
end.
}
\description{
Define a time range from a character string in suitable
format. The format is \eqn{start}\code{_}\eqn{end} where
\eqn{start} and \eqn{end} define the beginning and the end of the
time range.
}
\examples{
timeRange("2020_2222")
## error: and underscore '_' is needed! 
try(timeRange("2020-2222"))
timeRange("2020-01-01_2222")
timeRange("2020-01-01_2222-09-21")

}
