% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVGEVBayesMethods.R
\docType{data}
\name{TVGEVBayesMethods}
\alias{TVGEVBayesMethods}
\alias{coef.TVGEVBayes}
\alias{vcov.TVGEVBayes}
\alias{summary.TVGEVBayes}
\alias{print.TVGEVBayes}
\title{Methods for the S3 Class \code{"TVGEVBayes"}}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
TVGEVBayesMethods

\method{coef}{TVGEVBayes}(object, type = c("MAP", "postMean"), ...)

\method{vcov}{TVGEVBayes}(object, ...)

\method{summary}{TVGEVBayes}(object, ...)

\method{print}{TVGEVBayes}(x, ...)
}
\arguments{
\item{object}{A \code{TVGEVBayes} object.}

\item{type}{The type of coefficient.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Methods for the S3 Class \code{"TVGEVBayes"}
}
\keyword{datasets}
