% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMAXandTOTS.R
\name{year2Date}
\alias{year2Date}
\title{Transform a Year into a Date}
\usage{
year2Date(str, type = c("start", "end"))
}
\arguments{
\item{str}{A character string.}

\item{type}{The type of date to return. This has only an effect
when only the year is given, as oppposed to a full date. Then,
when \code{type} is \code{"start"}, the character \code{str} is
expected to describe the beginning of a time range and the
returned date will be the beginning of the year. When instead
\code{type} is \code{"end"}, the character \code{str} is expected
to describle the end of a time range and the returned date will be
the beginning of the next year. See \bold{Examples}.}
}
\value{
An object of class \code{"Date"} representing either the
beginning or the end of a time range.
}
\description{
Transform a year given as a string into a date corresponding
either the beginning of the year ot the begining of the following
year.
}
\section{Caution}{
 This function is not expected to work when
\code{str} has length \code{> 1}.
}

\examples{
year2Date("2021")
year2Date("2021", type = "end")

}
