% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beverstan-package.R
\docType{package}
\name{beverstan-package}
\alias{beverstan-package}
\alias{beverstan}
\title{The 'beverstan' package.}
\description{
Bayesian inference for some Extreme-Value models such as
Time-Varying GEV models for block maxima.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
